update dash core to 12.2

+++ src/crypto/ripemd160.h
@@ -14,7 +14,7 @@ class CRIPEMD160
 private:
     uint32_t s[5];
     unsigned char buf[64];
-    uint64_t bytes;
+    size_t bytes;
 
 public:
     static const size_t OUTPUT_SIZE = 20;
